+++ example/lib/app/settings/controller/app_settings_controller.dart
@@ -12,6 +12,15 @@ class AppSettingsController extends StateXController
 
   static AppSettingsController? _this;
 
+  /// Called when it's [StateX] object is itself disposed of.
+  @override
+  void dispose() {
+    // Good practice to nullify static instance reference.
+    // Flutter's garbage collection does its best, but why not if no longer used
+    _this = null;
+    super.dispose();
+  }
+
   ///
   @override
   Future<bool> initAsync() async {