+++ example/lib/home/controller/yet_another_controller.dart
@@ -12,4 +12,13 @@ class YetAnotherController extends StateXController
   factory YetAnotherController() => _this ??= YetAnotherController._();
   YetAnotherController._() : super();
   static YetAnotherController? _this;
+
+  /// Called when it's [StateX] object is itself disposed of.
+  @override
+  void dispose() {
+    // Good practice to nullify static instance reference.
+    // Flutter's garbage collection does its best, but why not if no longer used
+    _this = null;
+    super.dispose();
+  }
 }